Add Kconfig options to create a disassembler listing
In the past, we added a hardcoded rule to the Makefile that generated a
disassembler listing for u-boot. Now we also want to have this for SPL
and probably even TPL. Add some clean Kconfig options to enable or
disable these disassembler listings.

Also use the correct infrastructure to have short and long command output
depending on variable V.