New version of addfsheader as a script, remove old C version
With i.MX8, F&S will make more use of F&S headers. The concept is
extended to also have a type, flags, padding for alignment and
optional parameters in the header. These paramaters can be a
description of up to 32 characters, up to 32 8-bit values, up to
16 16-bit values, up to 8 32-bit values or up to 4 64-bit values.
Also an arbitrary combination of parameters is possible as long as
they do not overlap.

Add a new version of addfsheader. This is now a bash script. Remove
the old C version.

The new version can also add several files to one header, which is
a kind of grouping command. Use

  ./scripts/addfsheader --help

to show the usage.

Change Makefile to use the new version when creating uboot.fs.
Also use the proper command infrastructure to have correct short
an long command output depending on variable V.