Add fsimage script
By using addfsheader, several images can be combined to a hierarchical
file structure. The new tool fsimage is the counterpart. You can list
the contents of such an F&S image and you can extract single parts
from it.

fsimage is also a bash script. Use

  ./scripts/fsimage.sh --help

to show the usage.