Add fs_image library to load and handle F&S images on fsimx8mm
Boot Process and F&S Images
---------------------------

NXP's way to start a board is to build a combined image with SPL,
ATF, TEE and U-Boot. The DRAM timings and the DRAM training firmware
are compiled into SPL, so they are part of SPL. This combined image
is then downloaded to NAND or eMMC on the board and can be booted.

This looks rather simple, but on a closer look it is not. For each
board type, each DRAM type (DDR3L, LPDDR4) and even each DRAM size,
a different combination is needed. So we would end up in a large set
of boot images, not only one for each board type, but one for each
single board variant.

This is rather unfortunate. We at F&S are used to having only one
boot image for a whole family of boards with the same CPU. We also
think that end customers should not need to worry about all that
low-level stuff like SPL, DRAM timings, DRAM training firmware, ATF,
TEE, board configurations, and so on. So our idea is to have all
these things already pre-installed on the board, so that the customer
only needs to download U-Boot and Linux, like before on our i.MX6
boards. And if we call this pre-installed image NBoot again, it is
actually completely the same.

So we need a way to build a different image with only the low-level
images. But we want to add several DRAM timings and several DRAM
training firmwares. And we want to add the board configurations for
all our boards. So that we can use the same NBoot image on all our
boards with the same CPU again.

Unfortunately with all this data, such an NBoot image is quite large,
several hundreds of kilobytes. But when SPL starts, only OCRAM and
TCM are available, no regular DRAM. NBoot will not fit into any of
these on-chip memories. So another constraint is that we need to load
the NBoot image in parts, storing those parts that we need at the
appropriate addresses and ignoring those parts that we do not need.
We call this intelligent download process "streaming" the image.

Having to analyze different types of images during this streaming
process would be impossible. So we also need an easy way to identify
the type of each following sub-image before it is actually being
loaded.

To fulfil all these constraints, we have defined our own image type,
the F&S image. An F&S image is simply any kind of data blob that is
preceeded by an F&S header that shows the size and also the type of
the data. Of course in our case, the data is the set of low-level
images.

We already have a tool called addfsheader that can create such F&S
images. Each call can combine a group of files and adds the F&S
header. By repeatedly doing this, the final NBoot image can be seen
as a hierachy of sub-images and each sub-image can easily be
identified by its F&S header. So for example all board configurations
are grouped in a BOARD-CONFIGS sub-image. The DRAM timings and DRAM
training firmwares are grouped in a DRAM-SETTINGS sub-image. ATF, TEE
and the DRAM-SETTINGS build the FIRMWARE sub-image. And so on. See
fs_image.c for a closer look at the NBoot image layout.

When the board boots up, SPL can load this image from NAND or eMMC
and configures interfaces like the serial debug port. Then it
initializes the DRAM, loads ATF and (optionally) TEE.

This is possible if the NBoot image is already installed. But for
a completely fresh board in the production, everything is empty.
So we also need a way to download this NBoot image from USB. Like
NXP we are using the Serial Download Protocol (SDP) for this.

So the final boot procedure is this:

Step 1:
If a boot device is present in the fuses, load the BOARD-ID and
BOARD-CFG from this device (NAND or eMMC).

Step 2:
Load the remaining NBoot part from the current boot device. In a
regular boot case, this is the same device, e.g. NAND or eMMC.
If not (for example if a new NBoot is going to be installed via
USB), go to Step 3. Actually only the FIRMWARE part of the NBoot
image is stored in NAND or eMMC. So SPL will load this part now.

Step 2a:
Activate the board configuration, e.g. set the serial debug port.

Step 2b:
Load the appropriate DRAM timing and the DRAM training firmware.
Activate DRAM (DDR).

Step 2c:
Load ATF.

Step 2d:
If requested, load TEE.

If all these steps succeed, go to Step 4.

Step 3:
If there was some error in the steps above or if the board is
currently booting from USB anyway, then download the NBoot image
by SDP from USB.

Step 3a:
If there is a BOARD-ID in the image, load it. This will override
the BOARD-ID that was probably loaded in Step 1.

Step 3b:
Load the BOARD-CFG. This will probably override the BOARD-CFG
loaded in Step 1. Activate it, e.g. set serial debug port.

Step 3c:
Load the appropriate DRAM timing and the DRAM training firmware.
Activate DRAM (DDR).

Step 3d:
Load ATF.

Step 3e:
If requested, load TEE.

Images may only contain parts of the necessary information. So repeat
the whole Step 3 until all sub-steps are successfully done.

Step 4:
Proceed with booting, e.g. load U-Boot.

About this commit
-----------------

This commit adds the fs_image library to handle such NBoot images.
There is a state machine that implements the streaming procedure
as layed out in the steps above. It can handle the different parts
of the image and processes them accordingly. So for example the
low-level board configuration and the DRAM initialization will be
done while downloading the NBoot image.

When loading the FIRMWARE part of the NBoot image in Steps 2b-2d
above, the same state machine can be used. So also NAND and eMMC
functions are provided by the library.

The library also provides some generic functions for loading
F&S images without the state machine, for example later in U-Boot.

(Remark: eMMC code is not fully functional yet.)