Add nboot build process for fsimx8mm
An NBOOT image on fsimx8mm consists of a set of sub-images.

- SPL
- BOARD-CONFIGS with different BOARD-CFGs
- FIRMWARE with DRAM-SETTINGS, ATF and optional TEE. DRAM-SETTINGS
  in turn consists of differen DRAM-TYPES. Each DRAM-TYPE in turn
  consists of a DRAM-FW and different DRAM-TIMINGs.
- EXTRA with some bash scripts to handle F&S images.

For a detailed description of the image layout, see fs_image.c.

The following changes are done to implement NBoot images.

- Add directory board/F+S/NXP-Firmware where all firmware files
  from NXP, that are not part of U-Boot, have to be stored in. At
  the moment these are the DRAM Training firmwares and the ATF image.
  Add a README.txt there so that the directory is not empty, which
  would cause problems with git.
- Add directory board/F+S/fsimx8mm/nboot, where all NBoot related
  files are located.
- Add target nboot to the global Makefile. This target builds stuff
  in the nboot directory.
- Add board/F+S/fsimx8mm/nboot/Makefile with build instructions for
  the NBoot image.
- Move the existing DRAM timings from board/F+S/fsimx8mm to the nboot
  directory. Rename them so that the "_timing" part is at the end of
  the filename. This is needed by the build process and is also
  ignored for the appropriate F&S headers. This helps to stay below
  the 32 characters limit for F&S header descriptions.
- Create a first set of board configurations, for example
  PCoreMX8MM-FERT3.110.dts and PCoreMX8MX-FERT5.120.dts.
- Add linker script board/F+S/fsimx8mm/nboot/dram-timings.lds needed
  to link the RAM timings.
- Add a rule to global Makefile to delete all .fs files in target
  clean.

To build the NBoot image, you first have to download the necessary
NXP firmware images and store them in board/F+S/NXP-Firmware.
IMPORTANT: These files must never be added to our GIT repository!

Then simply call

  make nboot

This will generate all F&S images in the nboot directory and finally
creates nboot-fsimx8mm.fs in U-Boot's top directory.

An NBoot image may optionally be preceeded by a BOARD-ID header. Then
the image is capable of installing (or replacing) the board type. Such
an image is required in the production. For example to create an image
with the BOARD-ID PCoreMX8MM-FERT3.110, call

  ./scripts/addfsheader.sh -t BOARD-ID -d PCoreMX8MM-FERT3.110 \
     nboot-fsimx8mm.fs > nboot.fs

The result is in nboot.fs.

To avoid files with F&S header being unintentionally committed to
git, add .fs files to .gitignore.