mxs_nand_spl: Do not stop when ONFI version is unknown
The driver stopped the initialization when it encountered an unkown
ONFI version. This is completely wrong. New versions typically only
add new functions, so even if the ONFI version is newer than the last
one known by the driver, the chip should still be usable with an older
ONFI version. So just output a warning in this case, set a high value
for the ONFI version so that comparisons will work and continue to bring
up the chip.