Allow using function fdt_print() from outside of cmd/fdt.c
The listing of a device tree with function fdt_print() can be very
useful also for other commands. F&S will use this in its upcoming
fsimage command. So remove the static keyword and provide a signature
in include/common.h. Add an fdt argument, so that it can work with
any FDT, not just the one referred to by variable working_fdt.