Improve fsimx8mm SPL settings
There is a reset pin for the MIPI-2-LVDS converter on some boards of the
fsimx8mm architecture. The pad must be initialized before the SPL trys
to communicate via I2C with the PMIC. This is necessary because in some
special cases the I2C bus is broken due to the level of the reset pin.