Improve fsimx8mm mac address setup
Our default F&S implementation is that the mac-address will be
automatically set by this priority list.
1.) Environment variable
2.) OTP Fuses
3.) CONFIG_ETHADDR_BASE

So in normal cases like fsimx6* we don´t set the environment variable in
the header file like nxp we set the variable dynamically (on the fly).
Add these implementation for fsimx8mm to support F&S syntax.

- Also improve config jumper setup on layout pcoremx8mx we always two
lans so change this.
- Remove unnecessary pad setings for FEC. The pads will be set by device
tree and must not be set by board file.