fsimx8mm: Add fuse subcommand for fsimage
Add an implementation for the command fsimage fuse. This command will
read the fuse settings from the BOARD-CFG and compares them with the
current fuse settings. If there is a change, blow the fuses, probably
after getting confirmation from the user. Reject invalid settings.

Add fuse settings to the current BOARD-CFGs.