fsimx8mm: Have separate nboot-info for NAND and MMC
Actually we need more settings in nboot-info than just the start and
size of the NBoot region.

- When saving NBoot in U-Boot with the fsimage command, we need to know
  where SPL is located in NAND and eMMC.
- When also U-Boot should be saved with this command, we need to know
  where U-Boot is located in NAND and eMMC.

When clearly thinking this through, SPL in fact only needs to know the
start offsets of the BOARD-CFG so that it can load it to OCRAM. Then it
can get all remaining offsets from the nboot-info there. U-Boot on the
other hand only needs to know where the BOARD-CFG is located in OCRAM
and again it can get all offsets from the nboot-info there. So moving
this information to nboot-info further reduces the dependencies between
SPL and U-Boot.

It is also not necessary to have all values twice, once for NAND and
once for MMC. One board can either boot from NAND, or from MMC, not
both. So have two different include files nboot-info-nand.dtsi and
nboot-info-mmc.dtsi and include only one of them in each BOARD-CFG
file. Then we can also drop the prefix "mmc-" and "nand-" from the
values which reduces the number of access functions in the code.

Also change board/F+S/fsimx8mm/nboot/Makefile accordingly.

Change fsimx8mm.h to match the new NAND layout. This also needs
some changes in the MTDPARTS settings. Add a similar eMMC layout.
Change all CONFIG values accordingly.