Also use indendation in scripts/fsimage.sh
The fsimage list command in U-Boot indents the image type/description
according to the nesting level. This helps to understand what images
are grouped together and what images are sub-images. So do the same in
the fsimage.sh bash script. To have enough room, drop the magic number
in the output, it did not show important information anyway.