Add CONFIG_SPL/TPL_AUTOBUILD, do not build SPL by default on fsimx8mm
On F&S boards based on i.MX8, SPL is needed as part of NBoot, but not
when building regular U-Boot. So it would be nice if SPL was not built
all the time when only U-Boot is needed.

By adding new CONFIG_SPL_AUTOBUILD, we can decide now, whether SPL is
automatically built as part of U-Boot in a regular "make" or "make all"
call, or not. This is set by default, so that all other configurations
behave like before. But we unset it in fsimx8mm_defconfig. in this case,
SPL is only built if calling specific SPL specific targets. Add a new
target spl for that purpose, so "make spl" will build SPL. For F&S, SPL
is automatically built when calling "make nboot".

For completeness, also add CONFIG_TPL_AUTOBUILD with the same purpose for
TPL, even if we do not need it ourselves.