Fix addfsheade.sh for yocto build
Yocto separates the source and the build directory so we need to use
absolute pathes to call the addfsheade.sh script.

Also, for the mktemp call we need to strip the possibly absolute path
form the file nam,e so we get a valid template name for the temp files.