Set boardinfo.name to PicoCoreMX8MX for boards with 8mx config
The new config mechanism matches the "board-name" entry with
boardinfo.name to find the right board_info. Without an entry for the
PicoCoreMX8MX this fails and the board type is set to unknown.

We decided to keep the name PicoCoreMX8MX, which will also be displayed
in the bootlog, because at the moment we can not fully hide the
differences between PicoCoreMX8MM and PicoCoreMX8MX.

This will make it more transparent to the customer which board he is
using.