Fix DSI_PHY_REF clock settings
Fix DSI_PHY_REF clock settings
Linux now configures the 24MHz Ref clock as source for the DSI_PHY_REF
clock and sets it to 12MHz.

We adapt this for uboot too.

The changes to fs-imx8mm.dtsi do not have any effect and are juste made
for the sake of completeness.