Improve fsimx8mm KSZ9893 Switch setup
For revision 1.10 of PCB PicoCoreMX8MX there are two strapping pins
which setup the Serial Interface Mode (MDIO, I2C, SPI). There is a HW
bug in revision 1.10 so we are in the wrong Serial Interface Mode and
are unable to communicate via I2C. Due to this issue we are
pre-configuring the strapping pin as GPIO and setup the correct value.
Afterwards we resetting the chip so the chip can parse the new strapping
configuration and we are able to communicate via I2C with the chip.
Later on the device tree will setup the correct mux option for the
strapping pin.

The setup_iomux_fec resets the PHY in board file but we only need this
for PCB PicoCoreMX8MX, because we have to setup the correct strapping
option before the network infrastructure will be loaded. For PCB
PicoCoreMX8MM the reset is done by the network infrastructure and is
setup by the device tree.