Add -ddr3l/-lpddr4 to device tree names for fsimx8mm/fsimx8mn
There are two different PCBs for PicoCoreMX8MM and PicoCoreMX8MN as well.
One is called PCoreMX8MM and uses LPDDR4 and one is called PCoreMX8Mx and
uses DDR3L memory. The hardware documentation uses the RAM type to
differentiate between the two types of PCBs. To be consistent with that,
the software should use the same mechanism.

Both PCBs can be used with either i.MX8M-Mini or i.MX8M-Nano SoC. This
results in the following four combinations (BOARD-CFG name prefix in
parenthesis).

  PCoreMX8MM with i.MX8M-Mini: picocoremx8mm-lpddr4.dts (PCoreMX8MM-*)
  PCoreMX8Mx with i.MX8M-Mini: picocoremx8mm-ddr3l.dts  (PCoreMX8MX-*)
  PCoreMX8MM with i.MX8M-Nano: picocoremx8mn-lpddr4.dts (PCoreMX8MN-*)
  PCoreMX8Mx with i.MX8M-Nano: picocoremx8mn-ddr3l.dts  (PCoreMX8MN-*)

This requires quite a lot of changes.

- Rename the according .dts files.
- Change arch/arm/dts/Makefile to use the new device tree names.
- Change the board-name in all BOARD-CFG files for fsimx8mm/fsimx8mn.
- Change the name entries in board_info[] in fsimx8mm.c/fsimx8mn.c.
- Change the name entreies in board_names[] in both spl.c files for
  fsimx8mm/fsimx8mn.
- Change fsimx8mm_defconfig and fsimx8mn_defconfig to use the new
  device tree names.
- Rename back the device-tree macros of picocoremx8mn-lpddr4.dts from
  CONFIG_PICOCOREMX8MN_* to CONFIG_PICOCOREMX8MM_* as this is the
  marking on the PCB. This simplifies changes that need to be done in
  each file (for Mini and Nano). And if we ever want to merge these two
  device trees in the future, it will also be easier.
- Rename the PCoreMX8MX-* BOARD-CFGs for fsimx8mn to PCoreMX8MN-*. This
  is the name that is stored in the WaWi data base.
- While we are at it, also fix the PMIC interrupt pin configuration in
  the device trees.

ATTENTION: Users need to update both, NBoot and U-Boot, to make this
change work!

Remark:
Using "PCoreMX8MX-" as BOARD-CFG prefix for the i.MX8M-Mini/DDR3L variant
was a bad choice as this does not match the WaWi data base. This makes
the name selection in the production significantly more complicated. It
should rather be "PCoreMX8MM-", too. Unfortunately, there are already too
many boards in the field, so it is too late to fix this. We have to live
with that. On the i.MX8M-Nano, it should be correct.